abstract out formatBytes
