+++ src/coreUtils/coreUtils.js
@@ -2,20 +2,6 @@ export function coreUtils() {
   const lodash = window._;
   const Flow = window.Flow;
   const moment = window.moment;
-  const formatBytes = bytes => {
-    const sizes = [
-      'Bytes',
-      'KB',
-      'MB',
-      'GB',
-      'TB',
-    ];
-    if (bytes === 0) {
-      return '0 Byte';
-    }
-    const i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)), 10);
-    return Math.round(bytes / Math.pow(1024, i), 2) + sizes[i];
-  };
   const padTime = n => `${(n < 10 ? '0' : '')}${n}`;
   const splitTime = s => {
     const ms = s % 1000;
@@ -66,7 +52,6 @@ export function coreUtils() {
     return code;
   };
   Flow.Util = {
-    formatBytes,
     formatMilliseconds,
     formatElapsedTime,
     formatClockTime,
